requirements.txt
fastapi
uvicorn
python-multipart
geopandas
shapely
pytesseract
google-generativeai
Pillow
torch
diff-plonk
